pyproject.toml
[tool.poetry]
name = "monitored_subprocess"
version = "0.1.0"
description = ""
authors = ["Venky Iyer <indigoviolet@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
attrs = "^20.3.0"

[tool.poetry.dev-dependencies]
mypy = "^0.790"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
